app/ProductAPI/requirements.txt
fastapi
uvicorn
cassandra-driver
python-jose[cryptography]
python-dotenv
requests
